Расскидал роли

+++ routes/api.php
@@ -27,12 +27,17 @@ Route::get('/book/{id}', [\App\Http\Controllers\Api\BooksController::class, 'sho
 Route::get('/allAuthors', [\App\Http\Controllers\Api\AuthorsController::class, 'index']);
 Route::get('/author/{id}', [\App\Http\Controllers\Api\AuthorsController::class, 'show']);
 
+Route::get('/users', [\App\Http\Controllers\Api\UserController::class, 'index']);
+Route::get('/user/{id}', [\App\Http\Controllers\Api\UserController::class, 'show']);
+
 //Auth route
 Route::group(['middleware' => ['auth:sanctum']], function () {
     Route::post('/logout', [\App\Http\Controllers\Api\AuthController::class, 'logout']);
     Route::get('/allProfiles', [\App\Http\Controllers\Api\BookInTheProfileController::class, 'index']);
     Route::get('/profile/{id}', [\App\Http\Controllers\Api\BookInTheProfileController::class, 'show']);
     Route::delete('/profile/deleteBook/{id}', [\App\Http\Controllers\Api\BookInTheProfileController::class, 'destroy']);
+
+    Route::put('/updateUser/{id}', [\App\Http\Controllers\Api\UserController::class, 'update']);
 
     Route::post('/estimationsAdd', [\App\Http\Controllers\Api\EstimationController::class, 'store']);
 
@@ -41,26 +46,21 @@ Route::group(['middleware' => ['auth:sanctum']], function () {
 
     Route::delete('/deleteUser/{id}', [\App\Http\Controllers\Api\UserController::class, 'destroy']);
 
-    Route::post('/createAuthor', [\App\Http\Controllers\Api\AuthorsController::class, 'store'])
-        ->middleware('isAdmin');
+    // Admin route
+    Route::post('/createAuthor', [\App\Http\Controllers\Api\AuthorsController::class, 'store'])->middleware('isAdmin');
+
+    Route::put('/updateAuthor/{id}', [\App\Http\Controllers\Api\AuthorsController::class, 'update'])->middleware('isAdmin');
+
+    Route::put('/updateBook/{id}', [\App\Http\Controllers\Api\BooksController::class, 'update'])->middleware('isAdmin');
+    Route::delete('/deleteBook/{id}', [\App\Http\Controllers\Api\BooksController::class, 'destroy'])->middleware('isAdmin');
+    Route::post('/createBook', [\App\Http\Controllers\Api\BooksController::class, 'store'])->middleware('isAdmin');
+
+    Route::delete('/deleteGenre/{id}', [\App\Http\Controllers\Api\GenreController::class, 'destroy'])->middleware('isAdmin');
+    Route::post('/createGenre', [\App\Http\Controllers\Api\GenreController::class, 'store'])->middleware('isAdmin');
+    Route::put('/genreUpdate/{id}', [\App\Http\Controllers\Api\GenreController::class, 'update'])->middleware('isAdmin');
 });
 
-// Admin route
-Route::put('/updateAuthor/{id}', [\App\Http\Controllers\Api\AuthorsController::class, 'update']);
-
-Route::get('/users', [\App\Http\Controllers\Api\UserController::class, 'index']);
-Route::get('/user/{id}', [\App\Http\Controllers\Api\UserController::class, 'show']);
-
-Route::put('/updateBook/{id}', [\App\Http\Controllers\Api\BooksController::class, 'update']);
-Route::delete('/deleteBook/{id}', [\App\Http\Controllers\Api\BooksController::class, 'destroy']);
-Route::post('/createBook', [\App\Http\Controllers\Api\BooksController::class, 'store']);
-
-Route::delete('/deleteGenre/{id}', [\App\Http\Controllers\Api\GenreController::class, 'destroy']);
-Route::post('/createGenre', [\App\Http\Controllers\Api\GenreController::class, 'store']);
-Route::put('/genreUpdate/{id}', [\App\Http\Controllers\Api\GenreController::class, 'update']);
-
 //еще думаю куда их
-Route::put('/updateUser/{id}', [\App\Http\Controllers\Api\UserController::class, 'update']);
 
 Route::get('/allFeedbacks', [\App\Http\Controllers\Api\FeedbackController::class, 'index']);
 Route::get('/feedback/{id}', [\App\Http\Controllers\Api\FeedbackController::class, 'show']);
